+++ pages/pagelisp.h
@@ -63,6 +63,7 @@ private slots:
     void on_streamButton_clicked();
     void on_recentFilterEdit_textChanged(const QString &arg1);
     void on_exampleFilterEdit_textChanged(const QString &arg1);
+    void on_infoButton_clicked();
 
 private:
     Ui::PageLisp *ui;